implemented useContext() && resloved child key={} error for /Search DropDown list
