+++ hw3/03-react/src/components/audio/styles/audio.css
@@ -17,7 +17,7 @@
   flex-direction: column;
   line-height: 1;
   font-family: var(--rhap_font-family);
-  width: 20%;
+  width: 30%;
   padding: 10px 15px;
   background-color: var(--rhap_background-color);
   box-shadow: 0 0 3px 0 rgba(0, 0, 0, 0.2);
