+++ hw3/03-react/src/main.jsx
@@ -2,6 +2,9 @@ import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App.jsx";
 
+// Data Provider
+import  {DataProvider}  from "./dataprovider/DataProvider";
+
 // Global CSS Page Styles
 import "./index.css";
 import "./styles/main.css";
@@ -9,6 +12,8 @@ import "./styles/navbar.css";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <BrowserRouter>
-    <App />
+    <DataProvider>
+      <App />
+    </DataProvider>
   </BrowserRouter>
 );
